content/base/test/test_bug461735.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=461735
-->
<head>
  <title>Test for Bug 461735</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=461735">Mozilla Bug 461735</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">
var errorFired = false;
window.onerror = function(message, uri, line) {
  is(message, "Script error.", "Should have empty error message");
  is(uri, "", "Should have empty error location URI");
  is(line, 0, "Shouldn't have a line here");
  errorFired = true;
}
</script>
<script src="bug461735-redirect1.sjs"></script>
<script>
  is(errorFired, true, "Should have error in redirected different origin script");
  errorFired = false;
</script>
<script type="application/javascript">
window.onerror = function(message, uri, line) {
  is(message, "c is not defined", "Should have correct error message");
  is(uri, "http://mochi.test:8888/tests/content/base/test/bug461735-post-redirect.js", "Unexpected error location URI");
  is(line, 3, "Should have a line here");
  errorFired = true;
}
</script>
<script src="bug461735-redirect2.sjs"></script>
<script>
  is(errorFired, true, "Should have error in same origin script");
</script>
</pre>
</body>
</html>
